requirements.txt
python-pptx==0.6.23
pillow>=10.0.0
pytesseract==0.3.10
opencv-python-headless>=4.10.0
numpy>=1.26.0
easyocr>=1.7.1
